requirements.txt
blinker==1.4
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.0.3
commonmark==0.9.1
Deprecated==1.3.1
distlib==0.4.0
filelock==3.20.0
Flask==2.2.3
Flask-Limiter==2.7.0
Flask-Login==0.5.0
Flask-WTF==1.2.2
gunicorn==20.1.0
idna==3.11
itsdangerous==2.1.0
Jinja2==3.1.6
limits==5.6.0
MarkupSafe==3.0.3
packaging==25.0
pip==25.3
platformdirs==4.5.0
psycopg2-binary==2.9.10
Pygments==2.19.2
python-dotenv==0.21.0
pytz==2023.3
requests==2.31.0
rich==12.6.0
setuptools==80.9.0
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
virtualenv==20.35.4
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.0
WTForms==3.2.1
